refactor codegenil tests
